+++ e2e/__tests__/customReporters.test.ts
@@ -18,51 +18,6 @@ beforeEach(() => cleanup(DIR));
 afterEach(() => cleanup(DIR));
 
 describe('Custom Reporters Integration', () => {
-  test('valid string format for adding reporters', () => {
-    const reporterConfig = {
-      reporters: ['<rootDir>/reporters/TestReporter.js'],
-    };
-
-    const {exitCode} = runJest('custom-reporters', [
-      '--config',
-      JSON.stringify(reporterConfig),
-      'add.test.js',
-    ]);
-
-    expect(exitCode).toBe(0);
-  });
-
-  test('valid array format for adding reporters', () => {
-    const reporterConfig = {
-      reporters: [
-        ['<rootDir>/reporters/TestReporter.js', {'Aaron Abramov': 'Awesome'}],
-      ],
-    };
-
-    const {exitCode, stdout} = runJest('custom-reporters', [
-      '--config',
-      JSON.stringify(reporterConfig),
-      'add.test.js',
-    ]);
-
-    expect(wrap(stdout)).toMatchSnapshot();
-    expect(exitCode).toBe(0);
-  });
-
-  test('invalid format for adding reporters', () => {
-    const reporterConfig = {
-      reporters: [[3243242]],
-    };
-
-    const {exitCode, stderr} = runJest('custom-reporters', [
-      '--config',
-      JSON.stringify(reporterConfig),
-      'add.test.js',
-    ]);
-
-    expect(exitCode).toBe(1);
-    expect(wrap(stderr)).toMatchSnapshot();
-  });
 
   test('default reporters enabled', () => {
     const {stderr, stdout, exitCode} = runJest('custom-reporters', [
